+++ src/messaging/mod.rs
@@ -143,7 +143,8 @@ impl MsgEnvelope {
                                 Ok(Section(payment.sender().into()))
                             }
                             // From `Payment` to `Metadata`.
-                            Some(Duty::Elder(ElderDuties::Payment)) => {
+                            Some(Duty::Elder(ElderDuties::Payment))
+                            | Some(Duty::Elder(ElderDuties::Metadata)) => {
                                 Ok(Section(cmd.dst_address()))
                             }
                             // As it reads; simply no such recipient ;)
@@ -155,7 +156,8 @@ impl MsgEnvelope {
                             // Accumulated at `Metadata`.
                             // I.e. this means we accumulated a section signature from `Payment` Elders.
                             // (this is done at `Metadata` Elders, and the accumulated section is added to most recent sender)
-                            Some(Duty::Elder(ElderDuties::Payment)) => {
+                            Some(Duty::Elder(ElderDuties::Payment))
+                            | Some(Duty::Elder(ElderDuties::Metadata)) => {
                                 Ok(Section(cmd.dst_address()))
                             }
                             // As it reads; simply no such recipient ;)