+++ src/messaging/network.rs
@@ -8,8 +8,8 @@
 // Software.
 
 use crate::{
-    Address, Blob, BlobAddress, DebitId, Error, PublicKey, ReplicaEvent, Result, Signature,
-    SignedTransfer, TransferAgreementProof, TransferValidated, XorName,
+    Address, Blob, BlobAddress, DebitId, Error, MessageId, MsgSender, PublicKey, ReplicaEvent,
+    Result, Signature, SignedTransfer, TransferAgreementProof, TransferValidated, XorName,
 };
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeSet;
@@ -54,16 +54,36 @@ pub enum NodeTransferCmd {
 
 ///
 #[allow(clippy::large_enum_variant)]
-#[derive(Debug, Hash, Eq, PartialEq, Clone, Serialize, Deserialize)]
+#[derive(Debug, Eq, PartialEq, Clone, Serialize, Deserialize)]
 pub enum NodeDataCmd {
-    ///
+    /// Duplicate a given chunk at another Adult
     DuplicateChunk {
-        ///
+        /// New Holder's name
         new_holder: XorName,
-        ///
+        /// Address of the blob to be duplicated
         address: BlobAddress,
-        ///
+        /// Current Holders
         fetch_from_holders: BTreeSet<XorName>,
+    },
+    /// Get Chunk from current holders for duplication
+    GetChunk {
+        /// New Holder's name
+        new_holder: XorName,
+        /// Address of the blob to be duplicated
+        address: BlobAddress,
+        /// Details of the section that authorised the duplication
+        section_authority: MsgSender,
+        /// Current Holders
+        fetch_from_holders: BTreeSet<XorName>,
+    },
+    /// Provide chunk to the new holder for duplication
+    GiveChunk {
+        /// Blob to be duplicated
+        blob: Blob,
+        /// Name of the new holder
+        new_holder: XorName,
+        /// MessageId of the Duplication Message
+        correlation_id: MessageId,
     },
 }
 
@@ -254,7 +274,18 @@ impl NodeCmd {
         use NodeTransferCmd::*;
         match self {
             System(NodeSystemCmd::RegisterWallet { section, .. }) => Section(*section),
-            Data(DuplicateChunk { new_holder, .. }) => Node(*new_holder),
+            Data(cmd) => match cmd {
+                DuplicateChunk { new_holder, .. } => Node(*new_holder),
+                GetChunk {
+                    fetch_from_holders, ..
+                } => Node(
+                    *fetch_from_holders
+                        .iter()
+                        .next()
+                        .unwrap_or(&XorName::random()),
+                ), // namesake
+                GiveChunk { new_holder, .. } => Node(*new_holder),
+            },
             Transfers(cmd) => match cmd {
                 ValidateSectionPayout(signed_debit) => Section(signed_debit.sender().into()),
                 RegisterSectionPayout(transfer_agreement) => {